fix(gateway): harden worker ingress provenance
